feat: add navbar responsiveness and fix hardcoded tw values
