+++ components/navbar/dropdown-menu.tsx
@@ -85,6 +85,7 @@ export function DropwdownMenu() {
           alt="Arrow dropdown"
           width={20}
           height={20}
+          className="max-lg:rotate-180"
         />
       </Menu.Button>
       <Transition
@@ -96,7 +97,7 @@ export function DropwdownMenu() {
         leaveFrom="transform opacity-100 scale-100"
         leaveTo="transform opacity-0 scale-95"
       >
-        <Menu.Items className="absolute -right-[30px] top-12 z-50 flex w-[342px] origin-top flex-col gap-6 overflow-hidden rounded-2xl border border-osmoverse-750 bg-[#160F38] p-8">
+        <Menu.Items className="absolute -right-4 flex w-[342px] origin-top flex-col gap-6 overflow-hidden rounded-2xl border border-osmoverse-750 bg-[#160F38] p-8 max-lg:bottom-16 lg:-right-7.5 lg:top-12">
           <DiamondGradient
             colors={["#FF8642", "#EB26EB", "#492CE1"]}
             height={450}
