+++ components/navbar/navigation-menu.tsx
@@ -1,0 +1,47 @@
+import { Fragment } from "react";
+import Image from "next/image";
+import Link from "next/link";
+import { DropwdownMenu } from "@/components/navbar/dropdown-menu";
+import { cn } from "@/lib/utils";
+
+const links = [
+  {
+    label: "Earn",
+    href: "#",
+  },
+  {
+    label: "Trade",
+    href: "#",
+  },
+  {
+    label: "Assets",
+    href: "#",
+  },
+];
+
+export default function NavigationMenu({ className }: { className?: string }) {
+  return (
+    <nav
+      className={cn(
+        "lg:has-[div[data-headlessui-state='open']]:bg-gradient-navbar-menu-open z-50 flex items-center gap-2 rounded-full border border-[#9084CF40] px-8 py-4 backdrop-blur-xl lg:px-4 lg:py-2",
+        className,
+      )}
+    >
+      {links.map(({ label, href }) => (
+        <Fragment key={label}>
+          <Link href={href} className="font-body1 text-body1">
+            {label}
+          </Link>
+          <Image
+            src={"/assets/icons/nav-links-divider.svg"}
+            alt="Nav Divider"
+            className="opacity-25"
+            width={4}
+            height={4}
+          />
+        </Fragment>
+      ))}
+      <DropwdownMenu />
+    </nav>
+  );
+}
